fix : 各合宿のサイドバーやヘッダーをcamplayoutに移動
